feat/#61: endMeeting, processAfterMeeting을 이벤트 기반 아키텍처로 리팩토링
- EndMeetingUseCase 인터페이스를 선언 및 구현
- 해당 use case는 회의가 끝나면 웹소켓 연결을 종료하고 세션을 삭제하는 기능을 담당
- 웹소켓 연결이 종료되면 WebSocketHandler에서 afterConnectionClosed 메서드에서 MeetingEndedEvent를 발행
- 이후 MeetingEndEventListener가 MeetingEndedEvent를 구독하여, 이벤트가 발생하면 processAfterMeeting 메서드가 실행
